feat: deploy-commands もローカル用エントリポイントを分離
- deployCommands.ts を共通ロジックに、エントリポイントを
  deployCommands.production.ts / deployCommands.local.ts に分離
- yarn deploy-commands:local でテスト用ギルドにのみコマンド登録
- ドキュメントにコマンドデプロイ手順を追加
