+++ src/scripts/deployCommands.ts
@@ -1,24 +1,21 @@
 import { REST, Routes, type SlashCommandBuilder } from "discord.js";
-import dotenv from "dotenv";
 import type Command from "../domain/types/command";
 import logger from "../infrastructure/logger";
 import registry from "../interfaces/discordjs/commands";
 
-function checkEnvVariables() {
-  dotenv.config();
+export function loadDeployEnv() {
   const token = process.env.TOKEN;
   const clientId = process.env.APP_ID;
   const guildId = process.env.GUILD_ID;
 
   if (!token || !clientId || !guildId) {
-    logger.error("Missing environment variables.");
+    logger.error("Missing environment variables: TOKEN, APP_ID, GUILD_ID");
     process.exit(1);
   }
   return { token, clientId, guildId };
 }
 
-// コマンドをデプロイする関数
-async function deployCommands(
+export async function deployCommands(
   token: string,
   clientId: string,
   guildId: string,
@@ -34,7 +31,7 @@ async function deployCommands(
   );
 
   logger.info(
-    `Started refreshing ${commandData.length} application (/) commands.`,
+    `Started refreshing ${commandData.length} application (/) commands for guild ${guildId}.`,
   );
   try {
     await rest.put(Routes.applicationGuildCommands(clientId, guildId), {
@@ -48,12 +45,3 @@ async function deployCommands(
     throw error;
   }
 }
-
-async function main() {
-  const { token, clientId, guildId } = checkEnvVariables();
-  await deployCommands(token, clientId, guildId);
-}
-
-main().catch((error) =>
-  logger.error("Failed to execute main function:", error),
-);
